attempt to fix progress bar
