+++ pkg/scanner/scanner.go
@@ -2,8 +2,10 @@ package scanner
 
 import (
 	"context"
+	"io"
 	"sync"
 
+	consoleHelpers "github.com/Checkmarx/kics/internal/console/helpers"
 	"github.com/Checkmarx/kics/pkg/kics"
 	"github.com/rs/zerolog/log"
 )
@@ -12,16 +14,21 @@ func StartScan(ctx context.Context, scanID string, noProgress bool, services []*
 	var wg sync.WaitGroup
 	wgDone := make(chan bool)
 	errCh := make(chan error)
+	currentQuery := make(chan float64, 1)
+	var wgProg sync.WaitGroup
+	startProgressBar(noProgress, 1500, &wgProg, currentQuery) // total 1500 just to see if it works
 
 	for _, service := range services {
 		wg.Add(1)
-		go service.StartScan(ctx, scanID, noProgress, errCh, &wg)
+		go service.StartScan(ctx, scanID, noProgress, errCh, &wg, currentQuery)
 	}
 
 	go func() {
 		defer func() {
+			close(currentQuery)
 			close(wgDone)
 		}()
+		wgProg.Wait()
 		wg.Wait()
 	}()
 
@@ -35,4 +42,13 @@ func StartScan(ctx context.Context, scanID string, noProgress bool, services []*
 		return err
 	}
 	return nil
+}
+
+func startProgressBar(hideProgress bool, total int, wg *sync.WaitGroup, progressChannel chan float64) {
+	wg.Add(1)
+	progressBar := consoleHelpers.NewProgressBar("Executing queries: ", 10, float64(total), progressChannel)
+	if hideProgress {
+		progressBar.Writer = io.Discard
+	}
+	go progressBar.Start(wg)
 }
